refactor(html): move pages
