+++ test/build.gradle.kts
@@ -20,6 +20,7 @@ kotlin {
 dependencies {
     api(project(":void-base"))
     api(project(":void-json"))
+    api(project(":void-html"))
     api(libs.org.jetbrains.kotlin.kotlin.stdlib)
     api(libs.org.jetbrains.kotlinx.kotlinx.coroutines.core)
     testImplementation(libs.org.jetbrains.kotlin.kotlin.test.junit5)
